requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
dataclasses-json==0.5.7
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
langchain==0.0.122
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.1
python-telegram-bot==20.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tenacity==8.2.0
tomli==2.0.1
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
